+++ src/main/java/blackjack/domain/player/Player.java
@@ -1,4 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.player;
+
+import blackjack.domain.card.Card;
 
 public interface Player {
     void drawCard(final Card card);
