+++ src/main/java/blackjack/domain/player/User.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
 public class User extends AbstractPlayer {
     public User() {
