+++ src/test/java/blackjack/domain/player/NameTest.java
@@ -1,7 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
 import static org.assertj.core.api.Assertions.*;
 
+import blackjack.domain.player.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
